# openapi/brokkr-v1.json
{
  "components": {
    "schemas": {
...
      "AuditLogListResponse": {
        "description": "Response structure for audit log list operations.",
        "properties": {
          "count": {
            "description": "Number of entries returned.",
            "minimum": 0,
            "type": "integer"
          },
          "limit": {
            "description": "Limit used for this query.",
            "format": "int64",
            "type": "integer"
          },
          "logs": {
            "description": "The audit log entries.",
            "items": {
              "$ref": "#/components/schemas/AuditLog"
            },
            "type": "array"
          },
          "offset": {
            "description": "Offset used for this query.",
            "format": "int64",
            "type": "integer"
          },
          "total": {
            "description": "Total count of matching entries (for pagination).",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "logs",
          "total",
          "count",
          "limit",
          "offset"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "admin_pak": {
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      },
      "agent_pak": {
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      },
      "generator_pak": {
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Brokkr broker: the central control-plane API server for distributing Kubernetes objects across clusters.",
    "license": {
      "name": "Elastic-2.0",
      "url": "https://www.elastic.co/licensing/elastic-license"
    },
    "title": "brokkr-broker",
    "version": "0.7.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/admin/audit-logs": {
      "get": {
        "description": "Returns audit log entries matching the specified filters, ordered by\ntimestamp descending (most recent first).\n\n# Authentication\n\nRequires admin PAK authentication.\n\n# Query Parameters\n\n- `actor_type`: Filter by actor type (admin, agent, generator, system).\n- `actor_id`: Filter by actor UUID.\n- `action`: Filter by action (exact match or prefix with *).\n- `resource_type`: Filter by resource type.\n- `resource_id`: Filter by resource UUID.\n- `from`: Filter by start time (inclusive).\n- `to`: Filter by end time (exclusive).\n- `limit`: Maximum results (default 100, max 1000).\n- `offset`: Number of results to skip.\n\n# Returns\n\n- `200 OK`: List of audit logs with pagination info.\n- `401 UNAUTHORIZED`: Missing or invalid authentication.\n- `403 FORBIDDEN`: Authenticated but not an admin.\n- `500 INTERNAL_SERVER_ERROR`: Database error.",
        "operationId": "list_audit_logs",
        "parameters": [
          {
            "description": "Filter by actor type (admin, agent, generator, system).",
            "example": "admin",
            "in": "query",
            "name": "actor_type",
            "required": false,
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Filter by actor ID.",
            "in": "query",
            "name": "actor_id",
            "required": false,
            "schema": {
              "format": "uuid",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Filter by action (exact match or prefix with *).",
            "example": "agent.*",
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Filter by resource type.",
            "example": "agent",
            "in": "query",
            "name": "resource_type",
            "required": false,
            "schema": {
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Filter by resource ID.",
            "in": "query",
            "name": "resource_id",
            "required": false,
            "schema": {
              "format": "uuid",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Filter by start time (inclusive, ISO 8601).",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Filter by end time (exclusive, ISO 8601).",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "description": "Maximum number of results (default 100, max 1000).",
            "example": 100,
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int64",
              "nullable": true,
              "type": "integer"
            }
          },
          {
            "description": "Number of results to skip.",
            "example": 0,
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "format": "int64",
              "nullable": true,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogListResponse"
                }
              }
            },
            "description": "Audit logs retrieved successfully"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid authentication"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authorized (admin only)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database error"
          }
        },
        "security": [
          {
            "admin_pak": []
          }
        ],
        "summary": "Lists audit logs with optional filtering and pagination.",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/config/reload": {
      "post": {
        "description": "This endpoint triggers a hot-reload of configurable settings without\nrequiring a broker restart. Only settings marked as \"dynamic\" can be\nreloaded; static settings (like database URL) require a restart.\n\n# Authentication\n\nRequires admin PAK authentication.\n\n# Returns\n\n- `200 OK`: Configuration reloaded successfully with list of changes.\n- `401 UNAUTHORIZED`: Missing or invalid authentication.\n- `403 FORBIDDEN`: Authenticated but not an admin.\n- `500 INTERNAL_SERVER_ERROR`: Failed to reload configuration.",
        "operationId": "reload_config",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigReloadResponse"
                }
              }
            },
            "description": "Configuration reloaded successfully"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid authentication"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authorized (admin only)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Failed to reload configuration"
          }
        },
        "security": [
          {
            "admin_pak": []
          }
        ],
        "summary": "Reloads the broker configuration from disk.",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/ws/connections": {
      "get": {
        "operationId": "list_ws_connections",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WsConnectionsResponse"
                }
              }
            },
            "description": "Current WS connections snapshot"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "admin_pak": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/agent-events": {
      "get": {
        "operationId": "list_agent_events",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentEvent"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List all agent events"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "admin_pak": []
          },
          {
            "agent_pak": []
          },
          {
            "generator_pak": []
          }
        ],
        "tags": [
          "agent-events"
        ]
      }
    },
    "/agent-events/{id}": {
      "get": {
        "operationId": "get_agent_event",
        "parameters": [
          {
            "description": "Agent event id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentEvent"
                }
              }
            },
            "description": "Get agent event by id"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent event not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "admin_pak": []
          },
          {
            "agent_pak": []
          },
          {
            "generator_pak": []
          }
        ],
        "tags": [
          "agent-events"
        ]
      }
    },
    "/agents": {
      "get": {
        "operationId": "list_agents",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Agent"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successfully retrieved agents"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "admin_pak": []
          }
        ],
        "tags": [
          "agents"
        ]
      },
      "post": {
        "operationId": "create_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewAgent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAgentResponse"
                }
              }
            },
            "description": "Successfully created agent"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "admin_pak": []
          }
        ],
        "tags": [
          "agents"
        ]
      }
    },
    "/agents/": {
      "get": {
        "operationId": "search_agent",
        "parameters": [
          {
            "description": "Name of the agent to search for",
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Name of the cluster to search in",
            "in": "query",
            "name": "cluster_name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            },
            "description": "Successfully found agent"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "admin_pak": []
          }
        ],
        "tags": [
          "agents"
        ]
      }
    },
    "/agents/{agent_id}/webhooks/pending": {
      "get": {
        "operationId": "get_pending_agent_webhooks",
        "parameters": [
          {
            "description": "Agent ID",
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PendingWebhookDelivery"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Pending webhook deliveries for this agent"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden - Agent access required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "admin_pak": []
          },
          {
            "agent_pak": []
          }
        ],
        "tags": [
          "webhooks"
        ]
      }
    },
    "/agents/{agent_id}/work-orders/pending": {
      "get": {
        "operationId": "list_pending_for_agent",
        "parameters": [
          {
            "description": "Agent ID",
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Filter by work type",
            "in": "query",
            "name": "work_type",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkOrder"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of pending work orders"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "admin_pak": []
          },
          {
            "agent_pak": []
          }
        ],
        "tags": [
          "work-orders"
        ]
      }
    },
...
    "/agents/{id}/diagnostics/pending": {
      "get": {
        "operationId": "get_pending_diagnostics",
        "parameters": [
          {
            "description": "ID of the agent",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DiagnosticRequest"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successfully retrieved pending diagnostics"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden - PAK does not have required rights"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "agent_pak": []
          }
        ],
        "tags": [
          "diagnostics"
        ]
      }
    },
...
    "/agents/{id}/heartbeat": {
      "post": {
        "operationId": "record_heartbeat",
        "parameters": [
          {
            "description": "ID of the agent",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully recorded agent heartbeat"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "agent_pak": []
          }
        ],
        "tags": [
          "agents"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}